+++ packages/middleware/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stoxifyorg/middleware
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @stoxifyorg/database@2.0.0
+
 ## 1.0.3
 
 ### Patch Changes